--- returnn/_frontend/__init__.py
@@ -1,4 +1,0 @@
-"""
-Internal frontend code.
-This is supposed to be backend independent.
-"""
