--- returnn/_internal_frontend_api.py
@@ -1,256 +1,0 @@
-"""
-Like :mod:`returnn.frontend_api` / :class:`returnn.frontend_api.Frontend`,
-but this is for RETURNN internal usage.
-"""
-
-from __future__ import annotations
-
-from typing import Generic, TypeVar, Any, Optional, Union, Type, Sequence, Tuple
-import contextlib
-from returnn.tensor import Tensor
-
-
-T = TypeVar("T")  # tf.Tensor, torch.Tensor or so
-
-
-class InternalFrontend(Generic[T]):
-    """
-    Private API, most of them operating on raw tensors
-    """
-
-    RawTensorType: Type[T]
-
-    @staticmethod
-    def get_dtype_name_raw(raw_tensor: T) -> str:
-        """
-        :return: dtype of raw tensor, as string
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def get_ndim_raw(raw_tensor: T) -> int:
-        """
-        :return: ndim of raw tensor. assumes it is known
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def get_shape_raw(raw_tensor: T) -> T:
-        """
-        :return: shape of raw tensor
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def get_shape_tuple_raw(raw_tensor: T) -> Tuple[Union[int, T]]:
-        """
-        :return: shape of raw tensor. assumes that ndim is known.
-            In eager frameworks, all dims are int.
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def get_known_shape_raw(raw_tensor: T) -> Tuple[Optional[int]]:
-        """
-        :return: shape of raw tensor, int for static known, None otherwise. assumes that ndim is known.
-            This will not create any ops.
-            In eager frameworks, all dims are known.
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def set_known_shape_raw(raw_tensor: T, shape: Tuple[Optional[int]]) -> None:
-        """
-        Sets the known shape of the raw tensor.
-        This is only supported in graph-based frameworks,
-        and just performs a check in eager frameworks.
-        """
-        # Default implementation for eager-based frameworks.
-        assert all(dim is not None for dim in shape)
-        rif = get_internal_frontend_by_tensor_type(type(raw_tensor))
-        existing_shape = rif.get_known_shape_raw(raw_tensor)
-        assert all(dim is not None for dim in existing_shape)
-        assert shape == existing_shape
-
-    @staticmethod
-    def fill_raw(shape: Union[Sequence[Union[int, T]], T], value: Union[Any, T]) -> T:
-        """
-        :param shape: shape
-        :param value: scalar value to fill
-        :return: raw tensor filled with value everywhere
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def compare_raw(a: T, kind: str, b: T) -> T:
-        """
-        :param a:
-        :param kind: "equal"|"==", "less"|"<", "less_equal"|"<=", "greater"|">", "greater_equal"|">=", "not_equal"|"!="
-        :param b:
-        :return: a `kind` b
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def combine_raw(a: T, kind: str, b: T) -> T:
-        """
-        :param a:
-        :param kind: "add"|"+", "sub"|"-", "mul"|"*", "truediv"|"/", "floordiv"|"//", "mod"|"%", "pow"|"**",
-            "max"|"maximum", "min"|"minimum", "logical_and", "logical_or", "squared_difference"
-        :param b:
-        :return: a `kind` b
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def reshape_raw(raw_tensor: T, shape: Union[Sequence[Union[int, T]], T]) -> T:
-        """
-        :param raw_tensor: raw tensor
-        :param shape: new shape
-        :return: reshaped raw tensor
-        """
-        raise NotImplementedError
-
-    @classmethod
-    def squeeze_raw(cls, raw_tensor: T, axes: Sequence[int]) -> T:
-        """
-        :param raw_tensor: raw tensor
-        :param axes: axes to squeeze
-        :return: squeezed raw tensor
-        """
-        # Default implementation using reshape_raw.
-        known_shape = cls.get_known_shape_raw(raw_tensor)
-        assert all([known_shape[axis] == 1 for axis in axes])
-        new_shape = [dim for a, dim in enumerate(cls.get_shape_tuple_raw(raw_tensor)) if a not in axes]
-        return cls.reshape_raw(raw_tensor, new_shape)
-
-    @staticmethod
-    def transpose_raw(raw_tensor: T, perm: Sequence[int]) -> T:
-        """
-        :param raw_tensor: raw tensor
-        :param perm: permutation
-        :return: transposed raw tensor
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def expand_dims_raw(raw_tensor: T, axis: int) -> T:
-        """
-        :param raw_tensor:
-        :param axis:
-        :return: raw tensor with new axis
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def expand_raw(raw_tensor: T, axis: int, dim: Union[int, T]) -> T:
-        """
-        :param raw_tensor:
-        :param axis: shape[axis] must be 1
-        :param dim: the new dim for shape[axis]
-        :return: shape[axis] expands to dim.
-            in PyTorch or other frameworks which support custom strides,
-            this is an efficient view and not a copy.
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def sequence_mask_raw(lengths: T, *, batch_major: bool = True) -> T:
-        """
-        Like tf.sequence_mask().
-
-        :param lengths: shape (batch,)
-        :param batch_major:
-        :return: tensor mask of shape (batch,maxlen) if batch_major else (maxlen,batch) of type bool
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    @contextlib.contextmanager
-    def name_scope_raw(name: str) -> Any:
-        """
-        Default implementation for eager-based frameworks:
-        Do nothing, tensors do not have a name.
-
-        :param name:
-        :return: context manager
-        """
-        # Default implementation for eager-based frameworks
-        pass  # nothing to do
-
-    @staticmethod
-    @contextlib.contextmanager
-    def control_dependencies_raw(dependencies: Sequence[Any]) -> Any:
-        """
-        Default implementation for eager-based frameworks:
-        Do nothing, we expect that the dependencies are already executed.
-
-        :param dependencies: raw tensors or ops
-        :return: context manager
-        """
-        # Default implementation for eager-based frameworks
-        yield
-
-    @staticmethod
-    def identity_with_control_dependencies_raw(raw_tensor: T, dependencies: Sequence[Any]) -> T:
-        """
-        Default implementation for eager-based frameworks:
-        Do nothing, we expect that the dependencies are already executed.
-
-        :param raw_tensor: raw tensor
-        :param dependencies: raw tensors or ops
-        :return: raw tensor
-        """
-        # Default implementation for eager-based frameworks
-        return raw_tensor
-
-    @staticmethod
-    def create_placeholder(tensor: Tensor) -> T:
-        """
-        :return: tf.placeholder in TF
-
-        This is really only for TensorFlow for the deprecated option auto_create_placeholders
-        and should not be used in other backends,
-        even in graph-based backends.
-        Rather, the logic to create placeholders should be done elsewhere.
-        """
-        raise Exception("create_placeholder not supported by backend")
-
-    @staticmethod
-    def runtime_sanity_checks(tensor: Tensor) -> Any:
-        """
-        Checks whether the tensor.raw_tensor is consistent with the tensor metadata.
-
-        In graph-based frameworks (TF graph), we return some operation here.
-        In eager frameworks, we would not return anything but instead directly perform the checks.
-        """
-        # By default, we do not do any checks. This is optional for the backend.
-        pass
-
-    @staticmethod
-    def is_valid_in_current_graph(tensor: Tensor) -> bool:
-        """
-        :return: whether the raw tensor is valid in the current graph.
-            In eager-mode frameworks, this is always true -- there is no graph.
-        """
-        return True
-
-    @staticmethod
-    def format_graph_output(raw_tensor: T, *, max_depth: Optional[int] = None) -> str:
-        """
-        :return: the computation graph leading to this tensor formatted.
-            In eager-mode frameworks, this is not supported and returns None.
-        """
-        return "<no-graph>"
-
-
-def get_internal_frontend_by_tensor_type(tensor_type: Type[T]) -> Type[InternalFrontend[T]]:
-    """
-    :param tensor_type: e.g. tf.Tensor
-    :return: internal frontend class
-    """
-    from .frontend_api import get_frontend_by_tensor_type
-
-    rf = get_frontend_by_tensor_type(tensor_type)
-    # noinspection PyProtectedMember
-    return rf._internal_frontend
