+++ returnn/frontend/__init__.py
@@ -1,0 +1,9 @@
+"""
+This package provides the frontend API.
+The frontend API supports multiple backends.
+https://github.com/rwth-i6/returnn/issues/1120
+
+The convention for the user is to do::
+
+    from returnn import frontend as rf
+"""
