+++ returnn/tensor/_tensor_extra.py
@@ -17,7 +17,7 @@ from returnn.util.basic import NotSpecified
 from .dim import Dim, batch_dim, VerifyOutShapeException
 import returnn.tensor.tensor as _t
 import returnn.tensor.marked_dim as _m
-import returnn.frontend_api as _frontend_api
+import returnn.frontend._api as _frontend_api
 import returnn._internal_frontend_api as _internal_frontend_api
 from ._tensor_mixin_base import _TensorMixinBase
