+++ returnn/tensor/_tensor_op_overloads.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Union
 if TYPE_CHECKING:
     from .tensor import Tensor  # just for type hints; otherwise use _t.Tensor
 
-import returnn.frontend_api as _frontend_api
+import returnn.frontend._api as _frontend_api
 from ._tensor_mixin_base import _TensorMixinBase
